pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-dealiasing-algorithms"
version = "0.2.0"
description = "Public radar Doppler velocity dealiasing reference docs and paper-based implementations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Contributors"}]
keywords = ["radar", "doppler", "dealiasing", "meteorology", "weather"]
dependencies = ["numpy>=1.24"]

[project.optional-dependencies]
examples = ["matplotlib>=3.7"]
dev = ["matplotlib>=3.7", "pytest>=8.0"]
realdata = ["arm_pyart>=2.2", "boto3>=1.34", "matplotlib>=3.7"]

[project.urls]
Repository = "https://example.invalid/open-dealiasing-algorithms"

[tool.setuptools]
packages = ["open_dealias"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
